update wikibot scripts
