+++ secretsfile
@@ -1,0 +1,2 @@
+("addonbot", "xxxxxxxxxx") # Insert your bot name and password
+
