+++ user-config.py
@@ -1,0 +1,5 @@
+family = 'xbmc'
+mylang = 'en'
+usernames['xbmc']['en'] = u'addonbot' # Insert your bot name
+console_encoding = 'utf-8'
+password_file = "secretsfile" # Insert path to password file
